+++ src/tools/storage.ts
@@ -4,7 +4,6 @@ import storage from 'node-persist';
 import { PLATFORM_NAME, PLUGIN_NAME } from '../homebridge/settings.js';
 
 export const STORAGE_KEY_USER_AUTH = 'auth';
-export const STORAGE_KEY_MQTT = 'mqtt';
 export const STORAGE_KEY_RECOVERY_RATES = 'rates';
 
 const STORAGE = new Map<string, string>();